requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
python-dotenv==1.0.0
pandas==2.2.2
openpyxl==3.1.2
razorpay==1.3.0
twilio==8.10.0
Flask-WTF==1.2.1
gunicorn==21.2.0